WIP parallel templating to verify new Template
