+++ src/Header.php
@@ -90,6 +90,7 @@ class Header extends View
         if (!$this->icon && !$this->elements) {
             $this->template->del('has_content');
             $this->template->set('title', $this->content);
+            $this->content = false;
         }
 
         parent::renderView();
