+++ src/ProgressBar.php
@@ -7,10 +7,7 @@ namespace atk4\ui;
 /**
  * Class implements ProgressBar.
  *
- * $bar = ProgressBar::addTo($app, [
- *  10,
- *  'label' => 'Processing files',
- *  ]);
+ * $bar = ProgressBar::addTo($app, [10, 'label' => 'Processing files']);
  */
 class ProgressBar extends View
 {
