+++ src/Template.php
@@ -3,20 +3,7 @@
 declare(strict_types=1);
 
 namespace atk4\ui;
-
-use atk4\data\Model;
-
-require_once __DIR__ . '/TemplateOld.php';
-if (class_exists(Template::class)) {
-    return;
-}
-
-/**
- * This class is a lightweight template engine. It's based around operating with
- * chunks of HTML code and the main aims are:
- *  - completely remove any logic from templates
- *  - speed up template parsing and manipulation speed.
- */
+//require_once __DIR__ . '/TemplateNew.php';class Template extends TemplateNew{}return;
 class Template implements \ArrayAccess
 {
     use \atk4\core\AppScopeTrait;
@@ -26,760 +13,186 @@ class Template implements \ArrayAccess
     /** @const string */
     public const TOP_TAG = '_top';
 
-    /** @var array */
-    private static $_filesCache = [];
-    /** @var array */
-    private static $_parseTemplateCache = [];
+    // must be declared because of ArrayAccess interface
+    public function offsetExists($name) { return $this->__call('offsetExists', func_get_args()); }
+    public function offsetGet($name) { return $this->__call('offsetGet', func_get_args()); }
+    public function offsetSet($name, $val) { $this->__call('offsetSet', func_get_args()); }
+    public function offsetUnset($name) { $this->__call('offsetUnset', func_get_args()); }
 
-    // {{{ Properties of a template
+    /** @var TemplateNew */
+    private $tnew;
+    /** @var TemplateOld */
+    private $told;
 
-    /**
-     * This is a parsed contents of the template organized inside an array. This
-     * structure makes it very simple to modify any part of the array.
-     *
-     * @var array
-     */
-    private $template;
+    public function __construct(string $template = '') {
+        $this->tnew = new TemplateNew();
+        $this->told = new TemplateOld();
 
-    /**
-     * List of all tags found inside template implementing faster access when manipulating the template.
-     *
-     * @var array
-     */
-    private $tagsIndex;
-
-    /**
-     * Contains information about where the template was loaded from.
-     *
-     * @var string
-     */
-    protected $source;
-
-    // }}}
-
-    // {{{ Core methods - initialization
-
-    // Template creation, interface functions
-
-    public function __construct(string $template = '')
-    {
         $this->loadTemplateFromString($template);
     }
 
-    private function exceptionAddMoreInfo(Exception $ex): Exception
-    {
-        $ex->addMoreInfo('tags', implode(', ', array_keys($this->tagsIndex)));
-        $ex->addMoreInfo('template', $this->template);
-        $ex->addMoreInfo('source', $this->source);
+    public function __clone() {
+        $this->tnew = clone $this->tnew;
+        $this->told = clone $this->told;
 
-        return $ex;
+        $this->diffAfter();
     }
 
-    // }}}
-
-    // {{{ Tag manipulation
-
-    /**
-     * This is a helper method which returns reference to element of template
-     * array referenced by a tag.
-     *
-     * Because there might be multiple tags and getTagRef is
-     * returning only one template, it will return the first
-     * occurrence:
-     *
-     * {greeting}hello{/},  {greeting}world{/}
-     *
-     * calling &getTagRef('greeting') will return reference to &'hello';
-     *
-     * @param int|string|null $ref Null to return the first tag
-     */
-    protected function &getTagRef(string $tag, $ref = null): array
-    {
-        if ($ref !== null) {
-            if (!isset($this->tagsIndex[$tag][$ref])) {
-                throw $this->exceptionAddMoreInfo(
-                    (new Exception('Tag not found in template'))
-                        ->addMoreInfo('tag', $tag . '#' . $ref)
-                );
-            }
-
-            $path = $this->tagsIndex[$tag][$ref];
-        } else {
-            if ($tag === self::TOP_TAG) {
-                return $this->template;
-            }
-
-            $tag = explode('#', $tag, 2)[0];
-            if (!isset($this->tagsIndex[$tag])) {
-                throw $this->exceptionAddMoreInfo(
-                    (new Exception('Tag not found in template'))
-                        ->addMoreInfo('tag', $tag)
-                );
-            }
-
-            $path = reset($this->tagsIndex[$tag]);
-        }
-
-        $vRef = &$this->template;
-        foreach ($path as $k) {
-            // we must always return an array
-            // very dirty hack for deleted tags because of https://github.com/atk4/ui/blob/ddcd70e37bd126cf998c21f4bb10a9aa85a59625/src/Template.php#L322
-            if (!isset($vRef[$k]) && is_array($vRef)) {
-                $vRef[$k] = [''];
-            }
-
-            $vRef = &$vRef[$k];
-        }
-
-        return $vRef;
+    private function getImpl(): object {
+        return $this->tnew;
     }
 
-    protected function getTagRefs(string $tag): array
-    {
-        if ($tag === self::TOP_TAG) {
-            return [&$this->template];
-        }
-
-        $tag = explode('#', $tag, 2)[0];
-        if (!isset($this->tagsIndex[$tag])) {
-            throw $this->exceptionAddMoreInfo(
-                (new Exception('Tag not found in template'))
-                    ->addMoreInfo('tag', $tag)
-            );
-        }
-
-        $vsRef = [];
-        foreach ($this->tagsIndex[$tag] as $ref => $ignore) {
-            $vsRef[$ref] = &$this->getTagRef($tag, $ref);
-        }
-
-        return $vsRef;
+    private function getImpl2(): object {
+        return $this->told;
     }
 
-    /**
-     * Checks if template has defined a specified tag.
-     * If multiple tags are passed in as array, then return true if all of them exist.
-     *
-     * @param string|array $tag
-     */
-    public function hasTag($tag): bool
+    public function __isset($name)
     {
-        // check if all tags exist
-        if (is_array($tag)) {
-            foreach ($tag as $t) {
-                if (!$this->hasTag($t)) {
-                    return false;
+        foreach (['res' => $this->getImpl(), 'res2' => $this->getImpl2()] as $resVarName => $impl) {
+            ${$resVarName} = \Closure::bind(static function () use ($impl, $name) {
+                return isset($impl->{$name});
+            }, null, $impl)();
+        }
+
+        $this->diffAfter();
+
+        return $res;
+    }
+
+    public function &__get($name)
+    {
+        foreach (['res' => $this->getImpl(), 'res2' => $this->getImpl2()] as $resVarName => $impl) {
+            ${$resVarName} = &\Closure::bind(static function &() use ($impl, $name) {
+                return $impl->{$name};
+            }, null, $impl)();
+        }
+
+        $this->diffAfter();
+
+        return $res;
+    }
+
+    public function __set($name, $value)
+    {
+        foreach ([$this->getImpl(), $this->getImpl2()] as $impl) {
+            \Closure::bind(static function () use ($impl, $name, $value) {
+                $impl->{$name} = $value;
+            }, null, $impl)();
+        }
+
+        $this->diffAfter();
+    }
+
+    public function __unset($name)
+    {
+        foreach ([$this->getImpl(), $this->getImpl2()] as $impl) {
+            \Closure::bind(static function () use ($impl, $name) {
+                unset($impl->{$name});
+            }, null, $impl)();
+        }
+
+        $this->diffAfter();
+    }
+
+    private static $cloningFlag = false;
+    public function &__call($name, $args)
+    {
+        // set app
+        foreach ([$this->getImpl(), $this->getImpl2()] as $impl) {
+            $impl->app = $this->app;
+        }
+
+        try {
+            foreach (['res' => $this->getImpl(), 'res2' => $this->getImpl2()] as $resVarName => $impl) {
+                ${$resVarName} = &\Closure::bind(static function &() use ($impl, $name, $args) {
+                    if ((new \ReflectionClass($impl))->getMethod($name)->returnsReference()) {
+                        return $impl->{$name}(...$args);
+                    }
+
+                    $v = $impl->{$name}(...$args);
+
+                    return $v;
+                }, null, $impl)();
+            }
+        } /*catch (\Throwable $e) {
+            throw new \Error(); // only for debug, we do not except any exception...
+        } */finally {
+            // unset app
+//            foreach ([$this->getImpl(), $this->getImpl2()] as $impl) {
+//                unset($impl->app);
+//            }
+        }
+
+        $this->diffAfter();
+
+        if (!self::$cloningFlag && $res instanceof TemplateNew || $res instanceof TemplateOld) {
+            self::$cloningFlag = true;
+            try {
+                $origRes = $res;
+                $res = new static();
+                $origProps = \Closure::bind(static function () use ($origRes) {
+                    return get_object_vars($origRes);
+                }, null, $origRes)();
+                foreach ($origProps as $k => $v) {
+                    $res->{$k} = $v;
                 }
-            }
-
-            return true;
-        }
-
-        // check if tag exist
-        $tag = explode('#', $tag, 2)[0];
-
-        return isset($this->tagsIndex[$tag]) || $tag === self::TOP_TAG;
-    }
-
-    /**
-     * Re-create tags index from scratch for the whole template.
-     */
-    public function rebuildTagsIndex(): void
-    {
-        $this->tagsIndex = [];
-        $this->rebuildTagsIndexRegion([], $this->template);
-    }
-
-    private function rebuildTagsIndexRegion(array $path, array $template): void
-    {
-        $path[] = null;
-
-        foreach ($template as $tag => $val) {
-            if (is_numeric($tag)) {
-                continue;
-            }
-
-            $path[array_key_last($path)] = $tag;
-
-            [$tag, $ref] = explode('#', $tag, 2);
-
-            $this->tagsIndex[$tag][$ref] = $path;
-            if (is_array($val)) {
-                $this->rebuildTagsIndexRegion($path, $val);
-            }
-        }
-    }
-
-    // }}}
-
-    // {{{ Manipulating contents of tags
-
-    /**
-     * Internal method for setting or appending content in $tag.
-     *
-     * @param string|array|Model $tag
-     * @param string             $value
-     * @param bool               $encode Should we HTML encode content
-     */
-    protected function _setOrAppend($tag, $value = null, $encode = true, $append = false, $throwIfNotFound = true): void
-    {
-        // check tag
-        if ($tag instanceof Model) {
-            $tag = $this->app->ui_persistence->typecastSaveRow($tag, $tag->get());
-        }
-
-        // $tag passed as associative array [tag=>value]
-        // in this case we don't throw exception if tags don't exist
-        if (is_array($tag) && $value === null) {
-            foreach ($tag as $t => $v) {
-                $this->_setOrAppend($t, $v, $encode, $append, false);
-            }
-
-            return;
-        }
-
-        if (!$tag) {
-            throw (new Exception('Tag is not set'))
-                ->addMoreInfo('tag', $tag)
-                ->addMoreInfo('value', $value);
-        }
-
-        // check value
-        if (!is_scalar($value) && $value !== null) {
-            throw (new Exception('Value should be scalar'))
-                ->addMoreInfo('tag', $tag)
-                ->addMoreInfo('value', $value);
-        }
-
-        // encode value
-        $value = (string) $value; // TODO, better to remove later in favor of strong string type
-
-        if ($encode) {
-            $value = htmlspecialchars($value, ENT_NOQUOTES, 'UTF-8');
-        }
-
-        // ignore not existent tags
-        if (!$throwIfNotFound && !$this->hasTag($tag)) {
-            return;
-        }
-
-        $emptyValFunc = function(&$ref) use (&$emptyValFunc) {
-            if (is_array($ref)) {
-                foreach ($ref as &$r) {
-                    $emptyValFunc($r);
-                }
-            } else {
-                $ref = '';
-            }
-        };
-
-        // set or append value
-        $template = $this->getTagRefs($tag);
-        foreach ($template as &$ref) {
-            if (!$append) {
-                $emptyValFunc($ref);
-            }
-            $ref[] = $value;
-        }
-
-        $this->rebuildTagsIndex();
-
-        return;
-    }
-
-    /**
-     * This function will replace region referred by $tag to a new content.
-     *
-     * If tag is found inside template several times, all occurrences are
-     * replaced.
-     *
-     * ALTERNATIVE USE(2) of this function is to pass associative array as
-     * a single argument. This will assign multiple tags with one call.
-     * Sample use is:
-     *
-     *  set($_GET);
-     *
-     * would read and set multiple region values from $_GET array.
-     *
-     * @param string|array|Model $tag
-     * @param string             $value
-     * @param bool               $encode Should we HTML encode content
-     *
-     * @return $this
-     */
-    public function set($tag, $value = null, $encode = true)
-    {
-        $this->_setOrAppend($tag, $value, $encode, false, true);
-
-        return $this;
-    }
-
-    /**
-     * Same as set(), but won't generate exception for non-existing
-     * $tag.
-     *
-     * @param string|array|Model $tag
-     * @param string             $value
-     *
-     * @return $this
-     */
-    public function trySet($tag, $value = null, bool $encode = true)
-    {
-        $this->_setOrAppend($tag, $value, $encode, false, false);
-
-        return $this;
-    }
-
-    /**
-     * Set value of a tag to a HTML content. The value is set without
-     * encoding, so you must be sure to sanitize.
-     *
-     * @param string|array|Model $tag
-     * @param string             $value
-     *
-     * @return $this
-     */
-    public function setHtml($tag, $value = null)
-    {
-        $this->_setOrAppend($tag, $value, false, false, true);
-
-        return $this;
-    }
-
-    /**
-     * See setHtml() but won't generate exception for non-existing
-     * $tag.
-     *
-     * @param string|array|Model $tag
-     * @param string             $value
-     *
-     * @return $this
-     */
-    public function trySetHtml($tag, $value = null)
-    {
-        $this->_setOrAppend($tag, $value, false, false, false);
-
-        return $this;
-    }
-
-    /**
-     * Add more content inside a tag.
-     *
-     * @param string|array|Model $tag
-     * @param string             $value
-     *
-     * @return $this
-     */
-    public function append($tag, $value, bool $encode = true)
-    {
-        $this->_setOrAppend($tag, $value, $encode, true, true);
-
-        return $this;
-    }
-
-    /**
-     * Same as append(), but won't generate exception for non-existing
-     * $tag.
-     *
-     * @param string|array|Model $tag
-     * @param string             $value
-     *
-     * @return $this
-     */
-    public function tryAppend($tag, $value, bool $encode = true)
-    {
-        $this->_setOrAppend($tag, $value, $encode, true, false);
-
-        return $this;
-    }
-
-    /**
-     * Add more content inside a tag. The content is appended without
-     * encoding, so you must be sure to sanitize.
-     *
-     * @param string|array|Model $tag
-     * @param string             $value
-     *
-     * @return $this
-     */
-    public function appendHtml($tag, $value)
-    {
-        $this->_setOrAppend($tag, $value, false, true, true);
-
-        return $this;
-    }
-
-    /**
-     * Same as append(), but won't generate exception for non-existing
-     * $tag.
-     *
-     * @param string|array|Model $tag
-     * @param string             $value
-     *
-     * @return $this
-     */
-    public function tryAppendHtml($tag, $value)
-    {
-        $this->_setOrAppend($tag, $value, false, true, false);
-
-        return $this;
-    }
-
-    /**
-     * Get value of the tag. Note that this may contain an array
-     * if tag contains a structure.
-     */
-    public function get(string $tag): array
-    {
-        return $this->getTagRef($tag);
-    }
-
-    /**
-     * Empty contents of specified region. If region contains sub-hierarchy,
-     * it will be also removed.
-     *
-     * IMPORTANT: This does not dispose of the tags which were previously
-     * inside the region. This causes some severe pitfalls for the users
-     * and ideally must be checked and proper errors must be generated.
-     *
-     * @param string|array $tag
-     *
-     * @return $this
-     */
-    public function del($tag)
-    {
-        if (is_array($tag)) {
-            foreach ($tag as $t) {
-                $this->tryDel($t);
-            }
-
-            return $this;
-        }
-
-        if ($tag === self::TOP_TAG) {
-            $this->loadTemplateFromString('');
-
-            return $this;
-        }
-
-
-        $emptyValFunc = function(&$ref) use (&$emptyValFunc) {
-            if (is_array($ref)) {
-                foreach ($ref as &$r) {
-                    $emptyValFunc($r);
-                }
-            } else {
-                $ref = '';
-            }
-        };
-
-        $template = $this->getTagRefs($tag);
-        foreach ($template as &$ref) {
-            $emptyValFunc($ref);
-        }
-
-        $this->rebuildTagsIndex();
-
-        return $this;
-    }
-
-    /**
-     * Similar to del() but won't throw exception if tag is not present.
-     *
-     * @param string|array $tag
-     *
-     * @return $this
-     */
-    public function tryDel($tag)
-    {
-        if (is_array($tag)) {
-            return $this->del($tag);
-        }
-
-        return $this->hasTag($tag) ? $this->del($tag) : $this;
-    }
-
-    // }}}
-
-    // {{{ ArrayAccess support
-    public function offsetExists($name)
-    {
-        return $this->hasTag($name);
-    }
-
-    public function offsetGet($name)
-    {
-        return $this->get($name);
-    }
-
-    public function offsetSet($name, $val)
-    {
-        $this->set($name, $val);
-    }
-
-    public function offsetUnset($name)
-    {
-        $this->del($name, null);
-    }
-
-    // }}}
-
-    // {{{ Template Manipulations
-
-    /**
-     * Executes call-back for each matching tag in the template.
-     *
-     * @param string|array $tag
-     *
-     * @return $this
-     */
-    public function eachTag($tag, \Closure $callable)
-    {
-        // array support
-        if (is_array($tag)) {
-            foreach ($tag as $t) {
-                $this->eachTag($t, $callable);
-            }
-
-            return $this;
-        }
-
-        foreach ($this->getTagRefs($tag) as $ref => &$vRef) {
-            $vRef = [call_user_func($callable, $this->renderRegion($vRef), $tag . '#' . $ref)];
-        }
-
-        return $this;
-    }
-
-    /**
-     * Creates a new template using portion of existing template.
-     *
-     * @return static
-     */
-    public function cloneRegion(string $tag)
-    {
-        $template = new static();
-        $template->app = $this->app;
-        if ($tag === self::TOP_TAG) {
-            $template->template = $this->template;
-            $template->source = $this->source;
-        } else {
-            $template->template = [self::TOP_TAG . '#0' => $this->get($tag)];
-            $template->source = 'clone of tag "' . $tag . '" from template "' . $this->source . '"';
-        }
-        $template->rebuildTagsIndex();
-
-        return $template;
-    }
-
-    // }}}
-
-    // {{{ Template Loading
-
-    /**
-     * Loads template from a specified file.
-     *
-     * @return $this
-     */
-    public function load(string $filename)
-    {
-        if ($this->tryLoad($filename) !== false) {
-            return $this;
-        }
-
-        throw (new Exception('Unable to read template from file'))
-            ->addMoreInfo('file', $filename);
-    }
-
-    /**
-     * Same as load(), but will not throw exception.
-     *
-     * @return $this|false
-     */
-    public function tryLoad(string $filename)
-    {
-        $filename = realpath($filename);
-        if (!isset(self::$_filesCache[$filename])) {
-            self::$_filesCache[$filename] = $filename !== false ? file_get_contents($filename) : false;
-        }
-
-        if (self::$_filesCache[$filename] === false) {
-            return false;
-        }
-
-        $str = preg_replace('~(?:\r\n?|\n)$~s', '', self::$_filesCache[$filename]); // always trim end NL
-        $this->loadTemplateFromString($str);
-        $this->source = 'loaded from file: ' . $filename;
-
-        return $this;
-    }
-
-    /**
-     * Initialize current template from the supplied string.
-     *
-     * @return $this
-     */
-    public function loadTemplateFromString(string $str)
-    {
-        $this->source = 'string: ' . $str;
-        $this->template = [];
-        $this->tags = [];
-        $this->tagCnt = [];
-
-        if ($str !== '') {
-            $this->parseTemplate($str);
-        }
-
-        return $this;
-    }
-
-    // }}}
-
-    // {{{ Template Parsing Engine
-
-    /**
-     * Used for adding unique tag alternatives. E.g. if your template has
-     * {$name}{$name}, then first would become 'name#0' and second 'name#1', but
-     * both would still respond to 'name' tag.
-     *
-     * @var array
-     */
-    private $tagCnt = [];
-
-    /**
-     * Register tag and return unique tag name.
-     */
-    protected function regTag(string $tag): string
-    {
-        if (!isset($this->tagCnt[$tag])) {
-            $this->tagCnt[$tag] = -1;
-        }
-        $nextIndex = ++$this->tagCnt[$tag];
-
-        return $tag . '#' . $nextIndex;
-    }
-
-    protected function parseTemplateTree(array &$inputReversed, string $openedTag = null): array
-    {
-        $prefix = array_pop($inputReversed);
-        $template = $prefix !== '' ? [$prefix] : [];
-
-        while (($tag = array_pop($inputReversed)) !== null) {
-            $firstChar = substr($tag, 0, 1);
-            if ($firstChar === '/') { // is closing tag
-                $tag = substr($tag, 1);
-                if ($openedTag === null
-                    || ($tag !== '' && $tag !== $openedTag)) {
-                    throw (new Exception('Template parse error: tag was not opened'))
-                        ->addMoreInfo('opened_tag', $openedTag)
-                        ->addMoreInfo('tag', $tag);
-                }
-
-                $openedTag = null;
-
-                break;
-            }
-
-            // is new/opening tag
-            $fullTag = $this->regTag($tag);
-            $template[$fullTag] = $this->parseTemplateTree($inputReversed, $tag);
-
-            $chunk = array_pop($inputReversed);
-            if ($chunk !== null && $chunk !== '') {
-                $template[] = $chunk;
+                $res->rebuildTagsIndex();
+            } finally {
+                self::$cloningFlag = false;
             }
         }
 
-        if ($openedTag !== null) {
-            throw (new Exception('Template parse error: tag is not closed'))
-                ->addMoreInfo('tag', $openedTag);
-        }
-
-        return $template;
+        return $res;
     }
 
-    /**
-     * Deploys parse recursion.
-     */
-    protected function parseTemplate(string $str): void
-    {
-        $cKey = $str;
-        if (!isset(self::$_parseTemplateCache[$cKey])) {
-            // expand self-closing tags {$tag} -> {tag}{/tag}
-            $str = preg_replace('~\{\$([-_:\w]+)\}~', '{\1}{/\1}', $str);
+    private function diffAfter() {
+        // compare state with new refactored template
+        $tnew = $this->tnew;
+        $tnewTemplate = \Closure::bind(function() use($tnew) { return $tnew->template; }, null, TemplateNew::class)();
+        $tnewTagsIndex = \Closure::bind(function() use($tnew) { return $tnew->tagsIndex; }, null, TemplateNew::class)();
+        $told = $this->told;
+        $toldTemplate = \Closure::bind(function() use($told) { return $told->template; }, null, TemplateOld::class)();
+        $toldTags = \Closure::bind(function() use($told) { return $told->tags; }, null, TemplateOld::class)();
 
-            $input = preg_split('~\{(/?[-_:\w]*)\}~', $str, -1, PREG_SPLIT_DELIM_CAPTURE);
-            $inputReversed = array_reverse($input); // reverse to allow to use fast array_pop()
-
-            $this->template = $this->parseTemplateTree($inputReversed);
-            $this->rebuildTagsIndex();
-
-            self::$_parseTemplateCache[$cKey] = [$this->template, $this->tagsIndex];
-            $this->template = null;
-            $this->tagsIndex = null;
-        }
-
-        [$this->template, $this->tagsIndex] = self::$_parseTemplateCache[$cKey];
-
-
-        // compare with old version
-        $tOld = new TemplateOld();
-        \Closure::bind(function() use($tOld, $cKey) {
-            $tOld->parseTemplate($cKey);
-        }, null, TemplateOld::class)();
-        $tOldTemplate = \Closure::bind(function() use($tOld) { return $tOld->template; }, null, TemplateOld::class)();
-        $tOldTags = \Closure::bind(function() use($tOld) { return $tOld->tags; }, null, TemplateOld::class)();
-        if ($this->template !== $tOldTemplate) {
-            print_r($tOldTemplate);
-            print_r($this->template);
-            exit;
-        }
-
-        $thisTags = [];
-        foreach ($this->tagsIndex as $tag => $vv) {
-            foreach ($vv as $ref => $path) {
-                $v = $this->template;
+        $tnewTags = [];
+        foreach ($tnewTagsIndex as $tag => $paths) {
+            foreach ($paths as $ref => $path) {
+                $v = $tnewTemplate;
                 foreach ($path as $p) {
+                    if (!isset($v[$p])) { // up-too-date template should never reach this code block
+                        continue 2;
+                    }
                     $v = $v[$p];
                 }
-                $thisTags[$tag][$ref] = $v;
+                $tnewTags[$tag][$ref] = $v;
             }
         }
-        if ($thisTags !== $tOldTags) {
-            print_r($tOldTags);
-            print_r($thisTags);
-            print_r($this->tagsIndex);
-            ob_end_flush();
-            ob_start(function() { exit; }); echo 'x'; ob_end_flush(); // exit immediatelly
-        }
-    }
 
-    // }}}
-
-    // {{{ Template Rendering
-
-    /**
-     * Render either a whole template or a specified region. Returns
-     * current contents of a template.
-     */
-    public function render(string $region = null): string
-    {
-        return $this->renderRegion($region !== null ? $this->get($region) : $this->template);
-    }
-
-    /**
-     * Walk through the template array collecting the values
-     * and returning them as a string.
-     */
-    protected function renderRegion(array $template): string
-    {
-        $res = [];
-        foreach ($template as $val) {
-            $res[] = is_array($val) ? $this->renderRegion($val) : $val;
+        foreach (debug_backtrace() as $f) { // getTagRef/getTagRefs test can not succeed
+            if (isset($f['class']) && $f['class'] === tests\TemplateTest::class) {
+                return;
+            }
         }
 
-        return implode('', $res);
-    }
+        if ($tnew->render() !== $told->render()) {
+            echo trim($tnew->render()) . "\n" . trim($told->render()) ."\n\n\n"  ;
 
-    // }}}
+            var_dump($tnew->render());
+            var_dump($told->render());
+
+            if ($tnewTemplate !== $toldTemplate) {
+                echo 'new template' . "\n";
+                print_r($tnewTemplate);
+                echo 'old template' . "\n";
+                print_r($toldTemplate);
+                throw new \Error();exit;//echo (new Exception ('Template mismatch'))->getHtml();ob_end_flush(); ob_start(function() { exit; }); echo 'x'; ob_end_flush(); // exit immediatelly
+            }
+
+            if ($tnewTags !== $toldTags) {
+                print_r($tnewTags);
+                print_r($toldTags);
+                throw new \Error();exit;//echo (new Exception ('Template mismatch'))->getHtml();ob_end_flush(); ob_start(function() { exit; }); echo 'x'; ob_end_flush(); // exit immediatelly
+            }
+        }
+    }
 }
